Packages repositories and entities refactor
