+++ src/main/java/com/alkemy/ong/data/repositories/CategoryRepository.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.data.repository;
+package com.alkemy.ong.data.repositories;
 
 import com.alkemy.ong.data.entities.CategoryEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
