+++ src/main/java/com/alkemy/ong/data/repositories/MembersRepository.java
@@ -1,6 +1,6 @@
-package com.alkemy.ong.data.members.repository;
+package com.alkemy.ong.data.repositories;
 
-import com.alkemy.ong.data.members.entity.MemberEntity;
+import com.alkemy.ong.data.entities.MemberEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
