+++ src/main/java/com/alkemy/ong/data/repositories/UserRepository.java
@@ -1,4 +1,4 @@
-package com.alkemy.ong.data.repository;
+package com.alkemy.ong.data.repositories;
 
 import com.alkemy.ong.data.entities.UserEntity;
 import org.springframework.data.jpa.repository.JpaRepository;
